fix errors in formatting and code placement
